+++ jest.setup.js
@@ -1,2 +1,4 @@
 /* eslint-env jest */
 jest.mock('wrap-ansi')
+
+require('css.escape') // Polyfill for CSS.escape